+++ apps/repository/js/annotation_repo/AnnotationRepoTable.tsx
@@ -79,6 +79,7 @@ export default class AnnotationRepoTable extends ExtendedReactTable<IProps, ISta
                                             <AnnotationPreview id={id}
                                                                text={annotation.text}
                                                                img={annotation.img}
+                                                               meta={annotation.meta}
                                                                created={annotation.created}/>
 
                                         );